chore: move fixture to where it is used
